+++ src/lib/waku_store/index.node.spec.ts
@@ -12,7 +12,7 @@ import {
 } from '../../test_utils';
 import { delay } from '../delay';
 import { Waku } from '../waku';
-import { WakuMessage } from '../waku_message';
+import { DecryptionMethod, WakuMessage } from '../waku_message';
 import {
   generatePrivateKey,
   generateSymmetricKey,
@@ -320,6 +320,122 @@ describe('Waku Store', () => {
     }
 
     waku2.store.addDecryptionKey(symKey);
+
+    dbg('Retrieve messages from store');
+    const messages = await waku2.store.queryHistory([], {
+      decryptionKeys: [privateKey],
+    });
+
+    expect(messages?.length).eq(3);
+    if (!messages) throw 'Length was tested';
+    expect(messages[0].payloadAsUtf8).to.eq(clearMessageText);
+    expect(messages[1].payloadAsUtf8).to.eq(encryptedSymmetricMessageText);
+    expect(messages[2].payloadAsUtf8).to.eq(encryptedAsymmetricMessageText);
+
+    await Promise.all([waku1.stop(), waku2.stop()]);
+  });
+
+  it('Retrieves history with asymmetric & symmetric encrypted messages on different content topics', async function () {
+    this.timeout(10_000);
+
+    nimWaku = new NimWaku(makeLogFileName(this));
+    await nimWaku.start({ persistMessages: true, lightpush: true });
+
+    const encryptedAsymmetricMessageText =
+      'This message is encrypted for me using asymmetric';
+    const encryptedAsymmetricContentTopic = '/test/1/asymmetric/proto';
+    const encryptedSymmetricMessageText =
+      'This message is encrypted for me using symmetric encryption';
+    const encryptedSymmetricContentTopic = '/test/1/symmetric/proto';
+    const clearMessageText =
+      'This is a clear text message for everyone to read';
+    const otherEncMessageText =
+      'This message is not for and I must not be able to read it';
+
+    const privateKey = generatePrivateKey();
+    const symKey = generateSymmetricKey();
+    const publicKey = getPublicKey(privateKey);
+
+    const [
+      encryptedAsymmetricMessage,
+      encryptedSymmetricMessage,
+      clearMessage,
+      otherEncMessage,
+    ] = await Promise.all([
+      WakuMessage.fromUtf8String(
+        encryptedAsymmetricMessageText,
+        encryptedAsymmetricContentTopic,
+        {
+          encPublicKey: publicKey,
+        }
+      ),
+      WakuMessage.fromUtf8String(
+        encryptedSymmetricMessageText,
+        encryptedSymmetricContentTopic,
+        {
+          symKey: symKey,
+        }
+      ),
+      WakuMessage.fromUtf8String(
+        clearMessageText,
+        encryptedAsymmetricContentTopic
+      ),
+      WakuMessage.fromUtf8String(
+        otherEncMessageText,
+        encryptedSymmetricContentTopic,
+        {
+          encPublicKey: getPublicKey(generatePrivateKey()),
+        }
+      ),
+    ]);
+
+    dbg('Messages have been encrypted');
+
+    const [waku1, waku2, nimWakuMultiaddr] = await Promise.all([
+      Waku.create({
+        staticNoiseKey: NOISE_KEY_1,
+        libp2p: { modules: { transport: [TCP] } },
+      }),
+      Waku.create({
+        staticNoiseKey: NOISE_KEY_2,
+        libp2p: { modules: { transport: [TCP] } },
+      }),
+      nimWaku.getMultiaddrWithId(),
+    ]);
+
+    dbg('Waku nodes created');
+
+    await Promise.all([
+      waku1.dial(nimWakuMultiaddr),
+      waku2.dial(nimWakuMultiaddr),
+    ]);
+
+    dbg('Waku nodes connected to nim Waku');
+
+    let lightPushPeers = waku1.lightPush.peers;
+    while (lightPushPeers.length == 0) {
+      await delay(100);
+      lightPushPeers = waku1.lightPush.peers;
+    }
+
+    dbg('Sending messages using light push');
+    await Promise.all([
+      waku1.lightPush.push(encryptedAsymmetricMessage),
+      waku1.lightPush.push(encryptedSymmetricMessage),
+      waku1.lightPush.push(otherEncMessage),
+      waku1.lightPush.push(clearMessage),
+    ]);
+
+    let storePeers = waku2.store.peers;
+    while (storePeers.length == 0) {
+      await delay(100);
+      storePeers = waku2.store.peers;
+    }
+
+    waku2.addDecryptionKey(symKey, {
+      contentTopics: [encryptedSymmetricContentTopic],
+      method: DecryptionMethod.Symmetric,
+    });
 
     dbg('Retrieve messages from store');
     const messages = await waku2.store.queryHistory([], {